+++ great_expectations/rule_based_profiler/estimators/exact_numeric_range_estimator.py
@@ -4,14 +4,14 @@ from typing import Dict, Optional
 import numpy as np
 
 from great_expectations.rule_based_profiler.domain import Domain
+from great_expectations.rule_based_profiler.estimators.numeric_range_estimation_result import (
+    NumericRangeEstimationResult,
+)
 from great_expectations.rule_based_profiler.estimators.numeric_range_estimator import (
     NumericRangeEstimator,
 )
 from great_expectations.rule_based_profiler.helpers.util import (
     build_numeric_range_estimation_result,
-)
-from great_expectations.rule_based_profiler.numeric_range_estimation_result import (
-    NumericRangeEstimationResult,
 )
 from great_expectations.rule_based_profiler.parameter_container import (
     ParameterContainer,
