+++ great_expectations/rule_based_profiler/estimators/numeric_range_estimator.py
@@ -6,7 +6,7 @@ import numpy as np
 
 from great_expectations.core.util import convert_to_json_serializable
 from great_expectations.rule_based_profiler.domain import Domain
-from great_expectations.rule_based_profiler.numeric_range_estimation_result import (
+from great_expectations.rule_based_profiler.estimators.numeric_range_estimation_result import (
     NumericRangeEstimationResult,
 )
 from great_expectations.rule_based_profiler.parameter_container import (
