+++ great_expectations/rule_based_profiler/helpers/util.py
@@ -24,7 +24,7 @@ from great_expectations.rule_based_profiler.domain import (
     INFERRED_SEMANTIC_TYPE_KEY,
     SemanticDomainTypes,
 )
-from great_expectations.rule_based_profiler.numeric_range_estimation_result import (
+from great_expectations.rule_based_profiler.estimators.numeric_range_estimation_result import (
     NUM_HISTOGRAM_BINS,
     NumericRangeEstimationResult,
 )
