+++ tests/rule_based_profiler/parameter_builder/test_parameter_computations.py
@@ -6,15 +6,14 @@ import pandas as pd
 from great_expectations.rule_based_profiler.estimators.bootstrap_numeric_range_estimator import (
     DEFAULT_BOOTSTRAP_NUM_RESAMPLES,
 )
+from great_expectations.rule_based_profiler.estimators.numeric_range_estimation_result import (
+    NumericRangeEstimationResult,
+)
 from great_expectations.rule_based_profiler.helpers.util import (
     compute_bootstrap_quantiles_point_estimate,
 )
 
 # Allowable tolerance for how closely a bootstrap method approximates the sample
-from great_expectations.rule_based_profiler.numeric_range_estimation_result import (
-    NumericRangeEstimationResult,
-)
-
 EFFICACY_TOLERANCE: float = 1.0e-2
 
 # Measure of "closeness" between "actual" and "desired" is computed as: atol + rtol * abs(desired)
